+++ test/models/event_log_final_error_test.rb
@@ -8,7 +8,7 @@ class EventLogFinalErrorTest < ActiveSupport::TestCase
     final_error_count = Database.select_one "SELECT COUNT(*) FROM #{MovexCdc::Application.config.db_user}.Event_Log_Final_Errors"
     final_error_count = max_error_count if final_error_count > max_error_count
     test_count = EventLogFinalError.final_error_count(max_count: max_error_count)
-    assert final_error_count == test_count, log_on_failure("Final error count should be #{test_count } but is #{final_error_count} instead")
+    assert final_error_count == test_count, log_on_failure("Final error count should be #{final_error_count } but is #{test_count} instead")
   end
 
   test "an_error_message" do